backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
tensorflow>=2.12.0
numpy>=1.23.0
Pillow>=10.0.0
